add polylm-13b, openbuddy-llama2-13b (#14)
* add openbuddy-llama-65b

* add polylm-13b

* add openbuddy-llama-65b sh

* not use fp16

+++ examples/pytorch/llm/src/llm_sft.py
@@ -37,7 +37,7 @@ class SftArguments:
     seed: int = 42
     resume_from_ckpt: Optional[str] = None
     dtype: str = field(
-        default='fp16', metadata={'choices': {'bf16', 'fp16', 'fp32'}})
+        default='bf16', metadata={'choices': {'bf16', 'fp16', 'fp32'}})
     ignore_args_error: bool = False  # True: notebook compatibility
 
     dataset: str = field(
@@ -55,7 +55,7 @@ class SftArguments:
     quantization_bit: Optional[int] = field(
         default=None, metadata={'choices': {4, 8}})
     bnb_4bit_comp_dtype: str = field(
-        default='fp16', metadata={'choices': {'fp16', 'bf16', 'fp32'}})
+        default='fp32', metadata={'choices': {'fp16', 'bf16', 'fp32'}})
     bnb_4bit_quant_type: str = field(
         default='nf4', metadata={'choices': {'fp4', 'nf4'}})
     bnb_4bit_use_double_quant: bool = True
